icon moved to components folder
