+++ src/components/Icon/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { ReactComponent as CSS } from '../assets/svg/css.svg';
-import { ReactComponent as Html } from '../assets/svg/html.svg';
-import { ReactComponent as Javascript } from '../assets/svg/javascript.svg';
+import { ReactComponent as CSS } from '../../assets/svg/css.svg';
+import { ReactComponent as Html } from '../../assets/svg/html.svg';
+import { ReactComponent as Javascript } from '../../assets/svg/javascript.svg';
 
 const iconTypes = {
   css: CSS,
